feat(machine-learning): added jsdoc comments and rename
