--- src/resources/MachineLearning/CCConfiguration/index.ts
@@ -1,2 +1,0 @@
-export * from './CCConfiguration';
-export * from './CCConfigurationInterfaces';
