+++ src/resources/MachineLearning/CaseClassificationConfiguration/index.ts
@@ -1,0 +1,2 @@
+export * from './CaseClassificationConfiguration';
+export * from './CaseClassificationConfigurationInterfaces';
